% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_fu_invitro.R
\name{calculate_fu_austin_hepatocytes}
\alias{calculate_fu_austin_hepatocytes}
\title{Austin algorithm for hepatocytes Fu calculation}
\usage{
calculate_fu_austin_hepatocytes(
  ionization,
  pKa,
  log_lipophilicity,
  conc_cell_millionml
)
}
\arguments{
\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pKa}{vector of length of 2 with pKa of the compound}

\item{log_lipophilicity}{LogP or LogMA of the compound}

\item{conc_cell_millionml}{concentration of hepatocytes (in million cells/mL)}

\item{ion_factor}{ionization factor}
}
\value{
fuInvitro
}
\description{
The Austin algorithm for calculating Fu in vitro for hepatocytes
}
