% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fu_plasma.R
\name{calculate_fu_from_Ks}
\alias{calculate_fu_from_Ks}
\title{Calculate Fu in plasma based on affinity to different components}
\usage{
calculate_fu_from_Ks(
  partition_albumin,
  partition_globulin,
  partition_membrane_lipids,
  partition_lipids,
  species
)
}
\arguments{
\item{partition_albumin}{partition to albumin (in L/kg)}

\item{partition_globulin}{partition to globulin (in L/kg)}

\item{partition_membrane_lipids}{partition to membrane lipids (in L/L)}

\item{partition_lipids}{LogP or LogMA of the compound}

\item{species}{species to be considered, now there is data for human, rat, dog, monkey, rabbit and mouse. Mind that this changes the composition in serum but the specific affinities to albumin still need to be used}
}
\value{
Fu_plasma value
}
\description{
Calculate the Fu in plasma based on the affinity to the different components: albumin,
globulin and membrane lipids such as the ones in lipoproteins
}
\examples{
calculate_fu_from_Ks("partition_albumin"=10^4.48,
               "partition_globulin"=10^2.16,
               "partition_membrane_lipids"=10^3.51,
               "partition_lipids"=100,
               "species"="human")
               
}
