% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_fu_invitro.R
\name{calculate_fu_halifax}
\alias{calculate_fu_halifax}
\title{Halifax algorithm for Fu calculation}
\usage{
calculate_fu_halifax(
  ionization,
  pka,
  ion_factor,
  log_lipophilicity,
  concentration_microsomes
)
}
\arguments{
\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pKa}{vector of length of 2 with pkA of the compound}

\item{X}{ionization factor}

\item{logLipo}{LogP or LogMA of the compound}

\item{cMicro_mgml}{concentration of microsomes mg/mL}
}
\value{
fuInvitro
}
\description{
The Halifax algorithm for calculating Fu in vitro
}
\examples{
calculate_fu_halifax("base",3,100,3,1)
calculate_fu_halifax(c("base",0),c(3,0),100,3,1)
}
