% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_fu_invitro.R
\name{calculate_fu_turner}
\alias{calculate_fu_turner}
\title{Turner algorithm for Fu calculation}
\usage{
calculate_fu_turner(ionization, pKa, log_lipophilicity, conc_mic_mgml)
}
\arguments{
\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pKa}{vector of length of 2 with pKa of the compound}

\item{log_lipophilicity}{LogP or LogMA of the compound}

\item{conc_mic_mgml}{concentration of microsomes (in mg/mL)}
}
\value{
fuInvitro
}
\description{
The Turner algorithm for calculating Fu in vitro
}
\examples{
calculate_fu_turner("acid",3,3,1)

}
