% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearance_IVIVE.R
\name{clearance_IVIVE}
\alias{clearance_IVIVE}
\title{Clearance IVIVE}
\usage{
clearance_IVIVE(
  typeValue,
  units,
  expData,
  typeSystem,
  partitionQSPR,
  logLipo,
  FBS,
  microplateType = NULL,
  ionization = NULL,
  volMedium = NULL,
  REF = NULL,
  hlcAt = NULL,
  pKa = NULL,
  fu = NULL,
  fu_hep = NULL,
  cMicro = NULL,
  cCells = NULL,
  BP = NULL,
  species = NULL
)
}
\arguments{
\item{typeValue}{describe what type of in vitro data for clearance it is}

\item{units}{this are the units of the value}

\item{expData}{depending on the type of value this can be a csv file or just one value, units are min or min-1}

\item{typeSystem}{microsomes or hepatocytes}

\item{partitionQSPR}{type of assumption used (Poulin and Theil, PK-Sim® Standard, Rodgers & Rowland, Schmidtt)}

\item{logLipo}{LogP or LogMA of the compound}

\item{FBS}{fraction of FBS or FCS used}

\item{microplateType}{type of microplate, 96, 24, 48-wells, etc.}

\item{ionization}{Vector of length 3 with ionization class, acid, neutral and base, if not input then it is c(0,0,0)}

\item{volMedium}{volume of medium in the well (in mL)}

\item{REF}{relative expression or activity factor}

\item{hlcAt}{Henry's Law Constant in atm/(m3*mol)}

\item{pKa}{pkA of the compound as a vector of length 3}

\item{fu}{Fraction unbound in plasma}

\item{fu_hep}{Fraction unbound in the in vitro hepatic system. if not known code will calculate it}

\item{cMicro}{concentration of microsome protein mg/mL}

\item{cCells}{concentration of hepatocytes used million cells/mL}

\item{BP}{Blood plasma ratio}

\item{species}{values can human and rat for now, defaulting to human}
}
\value{
Specific clearance parameter to plug in PK-Sim
}
\description{
IVIVE for clearance
}
\examples{

expData <- read.csv("tests/Clearance_example.csv", header = TRUE)
library(ggplot2)
clearance_IVIVE(
  typeValue = "decay experimentalcurve", expData = expData,
  partitionQSPR = "All Schmitt",
  logLipo = 5,
  ionization = c("acid", 0, 0),
  typeSystem = "hepatocytes",
  FBS = 0.0, pKa = c(3, 0, 0),
  hlcAt = 1E-6,
  microplateType = 96,
  volMedium = 0.2,
  cCells = 0.02
)

}
