% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fu_plasma.R
\name{correct_fu_pearce}
\alias{correct_fu_pearce}
\title{Correct Fu based on Pearce correction factor}
\usage{
correct_fu_pearce(fraction_unbound, log_lipophilicity)
}
\arguments{
\item{fraction_unbound}{Fraction unbound in plasma}

\item{log_lipophilicity}{LogP or LogMA of the compound}
}
\value{
Corrected Fu_plasma value
}
\description{
Corrects the Fu based on Pearce correction factor for neutral lipids in plasma
}
\examples{
correct_fu_pearce(fraction_unbound=0.2, log_lipophilicity=4)
}
