% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_clearance_from_curve.R
\name{fit_clearance_from_curve}
\alias{fit_clearance_from_curve}
\title{function to determine clearance from experimental curve}
\usage{
fit_clearance_from_curve(expData_tmin_cuM, verbose = FALSE)
}
\arguments{
\item{expData_tmin_cuM}{table with first column as time in minutes and second column as concentration in uM}

\item{verbose}{if TRUE, print the inputs and the resulting kcat}
}
\value{
kcat in per min, this is not clearance ready for pksim
}
\description{
function to determine clearance from experimental curve
}
\examples{
exp_path<-system.file("extdata","clearance.csv",package="esqIVIVE")
expData<-read.csv(exp_path)
#see that the first column is time and second is concentration
expData

fit_clearance_from_curve(expData)


}
