% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mm_from_curve.R
\name{fit_mm_from_curve}
\alias{fit_mm_from_curve}
\title{Get michaelis-menten  parameters form experimental curves}
\usage{
fit_mm_from_curve(experimental_conc_velocity, verbose = FALSE)
}
\arguments{
\item{experimental_conc_velocity}{is a experimental curve with concentration in the first column and velocity in the second column}

\item{verbose}{if TRUE, print the inputs and the resulting Km/Vmax}
}
\value{
fitresults_vmax_km
}
\description{
Function to derive micahelis-menten form raw data
}
\examples{
mm_curve_path<-system.file("extdata","michaelis_menten_curve.csv",package="esqIVIVE")
mm_curve<-read.csv(mm_curve_path)
fit_mm_from_curve(mm_curve)

}
