% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Fu_inc_mic.R
\name{FractionUnbound}
\alias{FractionUnbound}
\title{getInVitroFractionUnbound}
\usage{
FractionUnbound(
  partitionQSPR,
  logLipo,
  ionization,
  typeSystem,
  FBS,
  microplateType,
  volMedium,
  pKa = NULL,
  hlcAt = NULL,
  fu = NULL,
  BP = NULL,
  cMicro = NULL,
  cCells = NULL
)
}
\arguments{
\item{partitionQSPR}{type of assumption used (Poulin and Theil, PK-Sim® Standard, Rodgers & Rowland, Schmidtt)}

\item{logLipo}{LogP or LogMA of the compound}

\item{ionization}{Vector of length 3 with ionization class, acid, neutral and base, if not input then it is c(0,0,0)}

\item{volMedium}{}

\item{pKa}{vector of pkA of the compound}

\item{hlcAt}{Henry's Law Constant in atm/(m3*mol)}

\item{fu}{In Vivo Fraction Unbound in plasma from literature}

\item{BP}{Blood plasma ratio}

\item{cMicro}{concentration of microsomes mg/mL}

\item{cCells}{concentration of cells million /mL}

\item{inVitroCompartment}{a list of values describing the in vitro compartment created by \code{getInVitroCompartment}}
}
\value{
fuInvitro and possible warning for evaporation
}
\description{
Compute the fraction unbound in vitro
}
\details{
mayeb consider to have average data..
}
\examples{
FractionUnbound(
  partitionQSPR = "All PK-Sim Standard", logLipo = 3, ionization = c("acid", 0, 0),
  typeSystem = "hepatocytes", FBS = 0, microplateType = 96,
  volMedium = 0.22, pKa = c(6, 0, 0), hlcAt = 1E-6, cCells = 2
)

}
