% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MM.R
\name{get_MM}
\alias{get_MM}
\title{Get michaelis-menten  parameters form experimental curves}
\usage{
get_MM(experimental_conc_velocity)
}
\arguments{
\item{experimental_conc_velocity}{is a experimental curve with concentration in the first column and velocity in the second column}
}
\value{
fitresults_vmax_km
}
\description{
Function to derive micahelis-menten form raw data
}
\examples{
mm_curve_path<-system.file("extdata","michaelis_menten_curve.csv","esqIVIVE")
mm_curve<-read.csv(mm_curve_path)
get_MM(mm_curve)

}
