% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_vitro_compartment.R
\name{getInVitroCompartment}
\alias{getInVitroCompartment}
\title{getInVitroCompartments}
\usage{
getInVitroCompartment(
  typeSystem,
  FBS_fraction,
  microplateType,
  volMedium_mL,
  cCells_Mml = NULL,
  cMicro_mgml = NULL
)
}
\arguments{
\item{typeSystem}{if system is hepatocytes or microsomes}

\item{FBS_fraction}{fraction of serum concentration, values can only go from 0-1}

\item{cCells_Mml}{cells concentration (in million/ml)}

\item{cMicro}{concentration of microsome protein mg/mL}

\item{microplateWells}{number of wells in the microplate}

\item{volMedium}{volume of medium in the well (in mL)}
}
\value{
a list of values representing the  different in vitro compartments, concentrations are given as fraction of volume
}
\description{
Generates a list of values describing an in vitro compartment
This function is used inside the Fraction unbound function
}
\examples{
getInVitroCompartment("hepatocytes", FBS_fraction=0.05, microplateType = 96, volMedium_mL = 0.15, cCells_Mml = 0.1)
getInVitroCompartment("microsomes", FBS_fraction=0, microplateType = 24, volMedium_mL = 0.5, cMicro_mgml = 1)

}
