% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_PartitionsC.R
\name{getInVitroFractionUnbound}
\alias{getInVitroFractionUnbound}
\title{getInVitroFractionUnbound}
\usage{
getInVitroFractionUnbound(
  assumptions,
  inVitroCompartment,
  LogP,
  hlcAt,
  MW,
  pKa,
  BC,
  ionization = "none",
  fu = NULL
)
}
\arguments{
\item{assumptions}{type of assumption used (Poulin and Theil, PK-Sim® Standard, Rodgers & Rowland, Schmidtt)}

\item{inVitroCompartment}{a list of values describing the in vitro compartment created by \code{getInVitroCompartment}}

\item{LogP}{LogP of the compound}

\item{hlcAt}{Henry's Law Constant in atm/(m3*mol)}

\item{MW}{Molecular Weight of the compound}

\item{pKa}{pkA of the compound}

\item{ionization}{type of ionization (acid, base, none)}

\item{fu}{In Vivo Fraction Unbound}
}
\description{
Compute the fraction unbound in vitro
}
\details{
Hepatocytes Assays: \eqn{fu_{in\ vitro}=\frac{1}{1 + C_{cells} \times 10^{0.4 \times logP - 1.38}}}
}
