% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_fu_invitro.R
\name{Halifax}
\alias{Halifax}
\title{Halifax algorithm for Fu calculation}
\usage{
Halifax(ionization, pKa, X, logLipo, cMicro_mgml)
}
\arguments{
\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pKa}{vector of length of 2 with pkA of the compound}

\item{X}{ionization factor}

\item{logLipo}{LogP or LogMA of the compound}

\item{cMicro_mgml}{concentration of microsomes mg/mL}
}
\value{
fuInvitro
}
\description{
The Halifax algorithm for calculating Fu in vitro
}
\examples{
Halifax("base",3,100,3,1)
Halifax(c("base",0),c(3,0),100,3,1)
}
