% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_vitro_compartments.R
\name{in_vitro_compartments}
\alias{in_vitro_compartments}
\title{in_vitro_compartments}
\usage{
in_vitro_compartments(
  typeSystem,
  FBS_fraction,
  microplateType,
  volMedium_mL,
  cCells_Mml = NULL,
  cMicro_mgml = NULL,
  verbose = FALSE
)
}
\arguments{
\item{typeSystem}{if system is hepatocytes or microsomes}

\item{FBS_fraction}{fraction of serum concentration, values can only go from 0-1}

\item{microplateType}{number of wells in the microplate}

\item{volMedium_mL}{volume of medium in the well (in mL)}

\item{cCells_Mml}{cells concentration (in million cells/mL)}

\item{cMicro_mgml}{concentration of microsome protein (in mg/mL)}

\item{verbose}{if TRUE, print the inputs and the resulting compartment list}
}
\value{
a list of values representing the  different in vitro compartments, concentrations are given as fraction of volume
}
\description{
Generates a list of values describing a liver in vitro compartment based on hepatocytes or microsomes
This function is used inside the Fraction unbound function but can also be used for general virtual hepatocyte systems
}
\examples{
in_vitro_compartments("hepatocytes", FBS_fraction=0.05, microplateType = 96,
                      volMedium_mL = 0.15, cCells_Mml = 0.1)
in_vitro_compartments("microsomes", FBS_fraction=0, microplateType = 24,
                      volMedium_mL = 0.5, cMicro_mgml = 1)

}
