% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_fu_invitro.R
\name{Kilford}
\alias{Kilford}
\title{Kilford algorithm for Fu calculation}
\usage{
Kilford(ionization, pKa, X, logLipo, cCells_Mml)
}
\arguments{
\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pKa}{vector of length of 2 with pkA of the compound}

\item{X}{ionization factor}

\item{logLipo}{LogP or LogMA of the compound}

\item{cCells_Mml}{concentration of hepatocytes million cells/mL}
}
\value{
fuInvitro
}
\description{
The Kilford algorithm for calculating Fu in vitro
}
