% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunction.R
\name{myFunction}
\alias{myFunction}
\title{myFunction}
\usage{
myFunction(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a numeric value
}
\description{
adds two numbers
}
\examples{
myFunction(x = 1, y = 2)
}
