% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pint_regressions.R
\name{pint_caco2_empir}
\alias{pint_caco2_empir}
\title{Function to convert from Papp to Pint transcellular permeability from a empirical regression}
\usage{
pint_caco2_empir(Papp_cms, verbose = FALSE)
}
\arguments{
\item{Papp_cms}{, permeability from Caco-2 in cms}

\item{verbose}{if TRUE, print the inputs and the resulting pint_cms}
}
\value{
pint in units cm/s
}
\description{
Function to convert from Papp to Pint transcellular permeability from a empirical regression
}
\examples{
pint_caco2_empir(Papp_cms=2.3E-6)

}
