% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pint_regressions.R
\name{pint_peff_empir}
\alias{pint_peff_empir}
\title{Function to convert from Papp to Pint transcellular permeability from a empirical regression}
\usage{
pint_peff_empir(Peff_cms, verbose = FALSE)
}
\arguments{
\item{Peff_cms}{Peff as obtained in vivo or QSARs in cms}

\item{verbose}{if TRUE, print the inputs and the resulting pint_cms}
}
\value{
pint in units cm/s
}
\description{
Function to convert from Papp to Pint transcellular permeability from a empirical regression
}
\examples{
pint_peff_empir(Peff_cms=2.3E-6)

}
