% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_fu_invitro.R
\name{Poulin}
\alias{Poulin}
\title{Poulin algorithm for Fu calculation}
\usage{
Poulin(ionization, pKa, X, Y, BP, fu, cCellNL, logLipo)
}
\arguments{
\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pKa}{vector of length of 2 with pkA of the compound}

\item{X}{ionization factor}

\item{Y}{intracellular ionization factor}

\item{BP}{blood plasma ratio}

\item{fu}{fraction unbound in plasma}

\item{cCellNL}{neutral lipid concentration}

\item{logLipo}{LogP or LogMA of the compound}
}
\value{
fuInvitro
}
\description{
The Poulin algorithm for calculating Fu in vitro
}
\examples{
Poulin("base",pKa=3,X=100,Y=120,BP=1,fu=0.2,cCellNL=0.03,logLipo=3)
Poulin("acid",pKa=3,X=100,cCellNL=0.03,logLipo=3)
}
