% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsars_plasma_components.R
\name{QSARs_plasma}
\alias{QSARs_plasma}
\title{QSARs for plasma components}
\usage{
QSARs_plasma(
  QSAR,
  logP,
  pKa,
  ionization,
  LFER_E = NULL,
  LFER_B = NULL,
  LFER_A = NULL,
  LFER_S = NULL,
  LFER_V = NULL
)
}
\arguments{
\item{QSAR}{type of QSAR, it can be logP based or PPLFER based (still sorting the ionization)}

\item{logP}{is the lipophilicity as given by logKow}

\item{pKa}{is a vector of length of 2}

\item{ionization}{is a vector of length of 2 which should indicate if chemical is neutral, acid basic.
There are spots for ionization in case chemical is zwitterion}

\item{LFER_E}{LFER E parameter}

\item{LFER_B}{LFER B parameter}

\item{LFER_A}{LFER A parameter}

\item{LFER_S}{LFER S parameter}

\item{LFER_V}{abraham volume}
}
\value{
kmemlip_LL, kalb_Lkg and kglob_Lkg
}
\description{
Collection of QSAR to obtain affinity to the different components in serum, membrane lipids (memlip)
albumin (alb) and globulin (glob)
}
\details{
For neutral chemicals with logP >4 use the logP QSAR
For acidic phenols, carboxylic acids, pyridine and amines you can use the PPLFER.
fup calculator (https://drumap.nibiohn.go.jp/fup/).
To DO:
PP-LFER QSARs.-need to check how ionization is considered
make documentation
}
\examples{

QSARs_plasma(QSAR="logP", logP=2, pKa=c(3,0), ionization=c("acid",0))

QSARs_plasma(QSAR="PPLFER", logP=2, pKa=c(3,0), ionization=c("acid",0), LFER_E=1, LFER_B=0, LFER_A=1.5, LFER_S=0.8, LFER_V=2)
}
