% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_fu_invitro.R
\name{Turner}
\alias{Turner}
\title{Turner algorithm for Fu calculation}
\usage{
Turner(ionization, pKa, X, logLipo, cMicro_mgml)
}
\arguments{
\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pKa}{vector of length of 2 with pkA of the compound}

\item{X}{ionization factor}

\item{logLipo}{LogP or LogMA of the compound}

\item{cMicro_mgml}{concentration of microsomes mg/mL}

\item{BP}{Blood plasma ratio, this parameter is needed for Rodgers and Rowland and Poulin method for basic chemicals}

\item{fu}{In Vivo Fraction Unbound in plasma from literature}

\item{cMicro}{concentration of microsomes mg/mL}
}
\value{
fuInvitro

fuInvitro
}
\description{
The different algorithms from literature to calculate Fu_mic

The Turner algorithm for calculating Fu in vitro
}
\details{
Algorithm4Fumic
}
\examples{
Turner("acid",3,100,3,1)
Halifax("base",3,100,3,1)
Halifax(c("base",0),c(3,0),100,3,1)
Austin("base",3,100,3,1)
Poulin("base",pKa=3,X=100,Y=120,BP=1,fu=0.2,cCellNL=0.03,logLipo=3,cMicro_mgml=1)
Poulin("acid",pKa=3,X=100,cCellNL=0.03,logLipo=3,cMicro_mgml=1)

}
