% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_fu_invitro.R
\name{calculate_fu_austin_hepatocytes}
\alias{calculate_fu_austin_hepatocytes}
\title{Austin algorithm for hepatocytes Fu calculation}
\usage{
calculate_fu_austin_hepatocytes(
  ionization,
  pka,
  ion_factor,
  log_lipophilicity,
  concentration_cells
)
}
\arguments{
\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pka}{vector of length of 2 with pkA of the compound}

\item{ion_factor}{ionization factor}

\item{log_lipophilicity}{LogP or LogMA of the compound}

\item{concentration_cells}{concentration of hepatocytes (in million cells/mL)}
}
\value{
fuInvitro
}
\description{
The Austin algorithm for calculating Fu in vitro for hepatocytes
}
