% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_fu_invitro.R
\name{calculate_fu_austin_microsomes}
\alias{calculate_fu_austin_microsomes}
\title{Austin algorithm for microsomes Fu calculation}
\usage{
calculate_fu_austin_microsomes(
  ionization,
  pKa,
  log_lipophilicity,
  conc_mic_mgml
)
}
\arguments{
\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pKa}{vector of length of 2 with pKa of the compound}

\item{log_lipophilicity}{LogP or LogMA of the compound}

\item{conc_mic_mgml}{concentration of microsomes (in mg/mL)}
}
\value{
fuInvitro
}
\description{
The Austin algorithm for calculating Fu in vitro for microsomes
}
\examples{
calculate_fu_austin_microsomes(c("base",0),c(8,0),3,1)

}
