% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms_fu_invitro.R
\name{calculate_fu_hep_austin}
\alias{calculate_fu_hep_austin}
\title{Austin algorithm for hepatocytes Fu calculation}
\usage{
calculate_fu_hep_austin(
  ionization,
  pKa,
  log_lipophilicity,
  conc_cell_millionml,
  verbose = FALSE
)
}
\arguments{
\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pKa}{vector of length of 2 with pKa of the compound}

\item{log_lipophilicity}{LogP or LogMA of the compound}

\item{conc_cell_millionml}{concentration of hepatocytes (in million cells/mL)}

\item{verbose}{if TRUE, print the inputs and the resulting fu_invitro}
}
\value{
fuInvitro
}
\description{
The Austin algorithm for calculating Fu in vitro for hepatocytes
}
\examples{
calculate_fu_hep_austin(ionization=c("base",0),pKa=c(3,0),log_lipophilicity=3,conc_cell_millionml=0.5)
}
