% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_fu_invitro.R
\name{calculate_fu_poulin}
\alias{calculate_fu_poulin}
\title{Poulin algorithm for Fu calculation}
\usage{
calculate_fu_poulin(
  ionization,
  pKa,
  blood_plasma = NULL,
  fraction_unbound = NULL,
  concentration_cell_neutral_lipids,
  log_lipophilicity
)
}
\arguments{
\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pKa}{vector of length of 2 with pKa of the compound}

\item{blood_plasma}{blood plasma ratio}

\item{fraction_unbound}{fraction unbound in plasma}

\item{concentration_cell_neutral_lipids}{neutral lipid concentration}

\item{log_lipophilicity}{LogP or LogMA of the compound}

\item{ion_factor_plasma}{ionization factor}

\item{ion_factor_cells}{intracellular ionization factor}
}
\value{
fuInvitro
}
\description{
The Poulin algorithm for calculating Fu in vitro
}
\examples{
calculate_fu_poulin(ionization=c("base",0),pKa=c(6,0),blood_plasma=1,fraction_unbound=0.2,concentration_cell_neutral_lipids=0.03,log_lipophilicity=3)
calculate_fu_poulin(ionization=c("neutral",0),pKa=c(0,0),concentration_cell_neutral_lipids=0.03,log_lipophilicity=3)
}
