% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_vitro_compartment.R
\name{calculate_in_vitro_compartments}
\alias{calculate_in_vitro_compartments}
\title{calculate_in_vitro_compartments}
\usage{
calculate_in_vitro_compartments(
  typeSystem,
  fetal_bovine_serum_fraction,
  microplateType,
  volMedium_mL,
  cCells_Mml = NULL,
  cMicro_mgml = NULL
)
}
\arguments{
\item{typeSystem}{if system is hepatocytes or microsomes}

\item{fetal_bovine_serum_fraction}{fraction of serum concentration, values can only go from 0-1}

\item{microplateType}{number of wells in the microplate}

\item{volMedium_mL}{volume of medium in the well (in mL)}

\item{cCells_Mml}{cells concentration (in million cells/mL)}

\item{cMicro_mgml}{concentration of microsome protein (in mg/mL)}
}
\value{
a list of values representing the  different in vitro compartments, concentrations are given as fraction of volume
}
\description{
Generates a list of values describing an in vitro compartment
This function is used inside the Fraction unbound function
}
\examples{
calculate_in_vitro_compartments("hepatocytes", fetal_bovine_serum_fraction=0.05, microplateType = 96, volMedium_mL = 0.15, cCells_Mml = 0.1)
calculate_in_vitro_compartments("microsomes", fetal_bovine_serum_fraction=0, microplateType = 24, volMedium_mL = 0.5, cMicro_mgml = 1)

}
