% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearance_ivive.R
\name{clearance_IVIVE}
\alias{clearance_IVIVE}
\title{IVIVE for clearance}
\usage{
clearance_IVIVE(
  typeValue,
  units,
  expData,
  typeSystem,
  fu_invitro = NULL,
  empirical_scalar = "No",
  tissue = NULL,
  species = NULL,
  volMedium = NULL,
  REF = NULL,
  cMicro_mgml = NULL,
  cCells_Mml = NULL,
  partitionQSPR = NULL,
  logLipo = NULL,
  ionization = NULL,
  FBS_fraction = NULL,
  microplateType = NULL,
  volMedium_mL = NULL,
  pKa = NULL,
  hlcAt_atmm3mol = NULL,
  fu = NULL,
  BP = NULL
)
}
\arguments{
\item{typeValue}{describe what type of in vitro data for clearance it is}

\item{units}{this are the units of the value. it does not matter typeValue is "decay experimentalcurve"}

\item{expData}{depending on the type of value this can be a csv file or just one value}

\item{typeSystem}{microsomes or hepatocytes}

\item{fu_invitro}{Fraction unbound in the in vitro hepatic system. if not known code will calculate it}

\item{empirical_scalar}{this is an option to include an extra empirical correction factor. Currently we are considering the scale factor of Wood 2017}

\item{tissue}{tissue of interest, since scaling factors are dependent on the tissue, will default to liver}

\item{species}{values can human and rat for now, defaulting to human}

\item{volMedium}{volume of medium in the well (in mL)}

\item{REF}{relative expression or activity factor}

\item{cMicro_mgml}{concentration of microsome protein mg/mL}

\item{cCells_Mml}{concentration of hepatocytes used million cells/mL}

\item{partitionQSPR}{type of assumption used for calculating fu_invitro (optional)}

\item{logLipo}{LogP or LogMA of the compound (optional)}

\item{ionization}{vector of length 2 with ionization class, acid, neutral and base (optional)}

\item{FBS_fraction}{fraction of serum concentration, values can only go from 0-1 (optional)}

\item{microplateType}{number of wells in the microplate (optional)}

\item{volMedium_mL}{volume of medium in the well (in mL) (optional)}

\item{pKa}{vector of length 2 with pKa values of the compound (optional)}

\item{hlcAt_atmm3mol}{Henry's Law Constant in atm/(m3*mol) (optional)}

\item{fu}{In Vivo Fraction Unbound in plasma from literature (optional)}

\item{BP}{Blood plasma ratio (optional)}
}
\value{
Specific clearance parameter to plug in PK-Sim
}
\description{
IVIVE for clearance
}
\examples{

expData<-read.csv("tests/testthat/data/clearance.csv",header=TRUE)
clearance_IVIVE(typeValue="decay experimentalcurve",typeSystem="hepatocytes",cCells_Mml=0.5,units="/minutes",
               expData=expData,fu_invitro=0.5,empirical_scalar="No")

# example hepatocytes
clearance_IVIVE(typeValue="in vitro clearance parameter",typeSystem="hepatocytes",species="human",units="mL/minutes/Millioncells",
expData=18.27,fu_invitro=0.5,cCells_Mml=0.5,empirical_scalar="No")

# example microsomes
clearance_IVIVE(typeValue="in vitro clearance parameter",typeSystem="microsomes",units="L/minutes/mg protein",
                expData=18.27,fu_invitro=0.5,cMicro_mgml=0.5,volMedium=0.5,empirical_scalar="No")

}
