% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fu_plasma.R
\name{correct_fu_Pearce}
\alias{correct_fu_Pearce}
\title{Correct Fu based on Pearce correction factor}
\usage{
correct_fu_Pearce(fu, lip_LL)
}
\arguments{
\item{fu}{Fraction unbound in plasma}

\item{lip_LL}{LogP or LogMA of the compound}
}
\value{
Corrected Fu_plasma value
}
\description{
Corrects the Fu based on Pearce correction factor for neutral lipids in plasma
}
\examples{
correct_fu_Pearce(fu=0.2, lip_LL=4)
}
