% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqIVIVE-package.R
\docType{package}
\name{esqIVIVE-package}
\alias{esqIVIVE}
\alias{esqIVIVE-package}
\title{esqIVIVE: IVIVE of ADME Parameters for PBK Models}
\description{
In vitro to in vivo extrapolation (IVIVE) of absorption, distribution, metabolism and excretion (ADME) parameters, to derive inputs for physiologically based kinetic (PBK) models. Predicts the fraction unbound in microsomal and hepatocyte incubations and in plasma, looks up measured fraction unbound values, fits substrate depletion and Michaelis-Menten curves, scales in vitro clearance to in vivo, and derives intestinal permeability from Caco-2 and effective permeability values. The functions are developed for use with the Open Systems Pharmacology Suite ('PK-Sim' and 'MoBi').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/esqLABS/esqIVIVE}
  \item \url{https://esqlabs.github.io/esqIVIVE/}
  \item Report bugs at \url{https://github.com/esqLABS/esqIVIVE/issues}
}

}
\author{
\strong{Maintainer}: Susana Proença \email{susana.proenca@esqlabs.com}

Authors:
\itemize{
  \item Susana Proença \email{susana.proenca@esqlabs.com}
}

Other contributors:
\itemize{
  \item Felix Mil \email{felix.mil@esqlabs.com} [contributor]
  \item ESQlabs GmbH [copyright holder, funder]
}

}
\keyword{internal}
