% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fu_inc_mic.R
\name{FractionUnbound}
\alias{FractionUnbound}
\title{getInVitroFractionUnbound}
\usage{
FractionUnbound(
  partitionQSPR,
  logLipo,
  ionization,
  typeSystem,
  FBS_fraction,
  microplateType,
  volMedium_mL,
  pKa = NULL,
  hlcAt_atmm3mol = NULL,
  fu = NULL,
  BP = NULL,
  cMicro_mgml = NULL,
  cCells_Mml = NULL
)
}
\arguments{
\item{partitionQSPR}{type of assumption used (Poulin and Theil, PK-Sim® Standard, Rodgers & Rowland, Schmidtt,
then from literature, Poulin, Turner, Austin and Halifax. See thevignette for more details)}

\item{logLipo}{LogP or LogMA of the compound}

\item{ionization}{Vector of length 2 with ionization class, acid, neutral and base, if not input then it is c(0,0)}

\item{pKa}{vector of length of 2 with pkA of the compound}

\item{hlcAt_atmm3mol}{Henry's Law Constant in atm/(m3*mol)}

\item{fu}{In Vivo Fraction Unbound in plasma from literature}

\item{BP}{Blood plasma ratio, this parameter is needed for Rodgers and Rowland and Poulin method for basic chemicals}

\item{cMicro_mgml}{concentration of microsomes mg/mL}

\item{cCells_Mml}{concentration of cells million /mL}

\item{inVitroCompartment}{a list of values describing the in vitro compartment created by \code{getInVitroCompartment}}

\item{volMedium}{}
}
\value{
fuInvitro and possible warning for evaporation
}
\description{
Compute the fraction unbound in vitro
}
\details{
mayeb consider to have average data..
}
\examples{
FractionUnbound(
 partitionQSPR = "All PK-Sim Standard", logLipo = 3, ionization = c("acid", 0),
 typeSystem = "hepatocytes", FBS_fraction = 0, microplateType = 96,
 volMedium_mL = 0.22, pKa = c(6, 0), hlcAt_atmm3mol = 1E-6, cCells_Mml = 2)

FractionUnbound(
 partitionQSPR = "Poulin and Theil + fu", logLipo = 3, ionization = c("acid", 0),
 typeSystem = "hepatocytes", FBS_fraction = 0, microplateType = 96,fu=0.01,BP=2,
 volMedium_mL = 0.22, pKa = c(6, 0), hlcAt_atmm3mol = 1E-6, cCells_Mml = 2)

FractionUnbound(
 partitionQSPR = "All Schmitt", logLipo = 0.42, ionization = c("acid", 0),
 typeSystem = "microsomes", FBS_fraction = 0, microplateType = 96,fu=0.2,BP=1,
 volMedium_mL = 0.22, pKa = c(6, 0), cMicro_mgml = 2)

}
