% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-vitro-compartment.R
\name{getInVitroCompartment}
\alias{getInVitroCompartment}
\title{getInVitroCompartments}
\usage{
getInVitroCompartment(cCells, cSerum, microplateType, volMedium)
}
\arguments{
\item{cCells}{cells concentration (in million/ml)}

\item{cSerum}{serum concentration (in ???)}

\item{volMedium}{volume of medium in the well (in mL)}

\item{microplateWells}{number of wells in the microplate}
}
\value{
a list of values representing
}
\description{
Generates a list of values describing an in vitro compartment
}
