% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_PartitionsC.R
\name{getInVitroFractionUnbound}
\alias{getInVitroFractionUnbound}
\title{getInVitroFractionUnbound}
\usage{
getInVitroFractionUnbound(
  LogP,
  hlcAt,
  MW,
  pKa,
  ionization,
  assumptions,
  fu,
  BC
)
}
\arguments{
\item{LogP}{LogP of the compound}

\item{hlcAt}{Henry's Law Constant in atm/(m3*mol)}

\item{MW}{Molecular Weight of the compound}

\item{pKa}{pkA of the compound}

\item{ionization}{type of ionization (acid, base, none)}

\item{assumptions}{type of assumption used (1: Poulin and Theil, 2: PK-Sim® Standard, 3: Rodgers & Rowland, 4: Schmidtt)}

\item{fu}{fraction unbound}

\item{BC}{}
}
\description{
Compute the fraction unbound in vitro
}
\details{
\deqn{fu_{in vitro}=1/(1 + C_{cells} \times 10^{0.4 \times logP - 1.38} )}
}
