% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion_factors.R
\name{ion_factors}
\alias{ion_factors}
\title{ion_factors}
\usage{
ion_factors(ionization, pKa)
}
\arguments{
\item{ionization}{vector of length 2 with ionization class, acid, neutral and base}

\item{pKa}{vector of length 2 with pKa values of the compound}
}
\value{
factors that can be used to calculate the fraction neutral or ionized in plasma and intracellularly
}
\description{
Function to calculate fraction unionized
If there are multiple pKas for acidity just use the lower value
If there are multiple pKas for basicity just use the lower value
Mind that pKb is not the same as pKa !
If you have pKb, just calculate pKa=14-pKb
}
\examples{
ion_factors(ionization=c("neutral",0),pKa<-c(0,0))
ion_factors(ionization=c("acid",0),pKa<-c(14,0))
ion_factors(ionization=c("base","acid"),pKa<-c(5,7))
}
