% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_clearance.R
\name{IVIVE_clearance}
\alias{IVIVE_clearance}
\title{IVIVE for clearance}
\usage{
IVIVE_clearance(
  typeValue,
  units,
  expData,
  typeSystem,
  fu_invitro = 1,
  empirical_scalar = "No",
  tissue = "liver",
  species = "human",
  volMedium_mL = 1,
  REF = 1,
  cProtein_mgml = NULL,
  cCells_Mml = NULL,
  verbose = FALSE
)
}
\arguments{
\item{typeValue}{what type of value it is: kcat (directly from in vitro), halfLife or invitro_clearance_parameter)}

\item{units}{this are the units of the value.
For kct\tabular{lll}{
   Cells \tab Subcellular \tab Generic \cr
   mL/minutes/millioncells \tab mL/minutes/mg protein \tab /minutes \cr
   uL/minutes/millioncells \tab uL/minutes/mg protein \tab /hours \cr
   L/minutes/millioncells \tab L/minutes/mg protein \tab /seconds \cr
   mL/hours/millioncells \tab mL/hours/mg protein \tab mL/minutes \cr
   uL/hours/millioncells \tab uL/hours/mg protein \tab uL/minutes \cr
   L/hours/millioncells \tab L/hours/mg protein \tab mL/seconds \cr
   mL/seconds/millioncells \tab mL/seconds/mg protein \tab uL/seconds \cr
   uL/seconds/millioncells \tab mL/seconds/mg protein \tab mL/hours \cr
   L/seconds/millioncells \tab uL/seconds/mg protein \tab uL/hours \cr
   mL/minutes/cell \tab L/seconds/mg protein \tab mL/minutes/kg \cr
   uL/minutes/cell \tab  \tab uL/minutes/kg \cr
   mL/hours/cell \tab  \tab mL/hours/kg \cr
   uL/hours/cell \tab  \tab uL/hours/kg \cr
   mL/seconds/cell \tab  \tab  \cr
   uL/seconds/cell \tab  \tab  \cr
}}

\item{expData}{experimental clearance value}

\item{typeSystem}{"cells", "microsomes" or "cytosolF" (cytosolic fraction). Cells are scaled with
the cells per gram organ (CellsGO), microsomes with the microsomal protein per gram organ (MicProtGO)
and the cytosolic fraction with the cytosolic protein per gram organ (CytosProtGO)}

\item{fu_invitro}{Fraction unbound in the in vitro hepatic system. if not known code will calculate it}

\item{empirical_scalar}{this is an option to include an extra empirical correction factor. Currently we are considering the scale factor of Wood 2017}

\item{tissue}{tissue of interest, since scaling factors are dependent on the tissue, will default to liver}

\item{species}{values can human and rat for now, defaulting to human}

\item{volMedium_mL}{volume of medium in the well (in mL)}

\item{REF}{relative expression or activity factor}

\item{cProtein_mgml}{concentration of subcellular (microsomal or cytosolic) protein (in mg/mL)}

\item{cCells_Mml}{concentration of cells used (in million cells/mL)}

\item{verbose}{if TRUE, print the inputs and the resulting clearance}
}
\value{
Specific clearance parameter (/min) to plug in PK-Sim
}
\description{
IVIVE for clearance based on different type of values
}
\examples{
# example cells (e.g. hepatocytes)
IVIVE_clearance(typeValue="invitro_clearance_parameter",typeSystem="cells",species="human",
units="mL/minutes/millioncells",expData=18.27,fu_invitro=0.5,cCells_Mml=0.5,empirical_scalar="No")

# if you dont specify some of the parameters they will be the default (example fu_in vitro=1)
IVIVE_clearance(typeValue="invitro_clearance_parameter",typeSystem="cells",
units="mL/minutes/millioncells",expData=18.27,cCells_Mml=0.5,verbose=TRUE)


# example microsomes
IVIVE_clearance(typeValue="invitro_clearance_parameter",typeSystem="microsomes",
                units="L/minutes/mg protein",expData=18.27,fu_invitro=0.5,cProtein_mgml=0.5,
                volMedium_mL=0.5,empirical_scalar="No")

# example cytosolic fraction
IVIVE_clearance(typeValue="invitro_clearance_parameter",typeSystem="cytosolF",
                units="mL/minutes/mg protein",expData=0.05,fu_invitro=0.8,cProtein_mgml=1)

}
