% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_clearance.R
\name{IVIVE_clearance}
\alias{IVIVE_clearance}
\title{IVIVE for clearance}
\usage{
IVIVE_clearance(
  typeValue,
  units,
  expData,
  typeSystem,
  fu_invitro = 1,
  empirical_scalar = "No",
  tissue = "liver",
  species = "human",
  volMedium_mL = 1,
  REF = 1,
  cMicro_mgml = NULL,
  cCells_Mml = NULL
)
}
\arguments{
\item{typeValue}{what type of value it is: kcat (directly form in vitro, half life, in vitro clearance parameter)}

\item{units}{this are the units of the value.}

\item{expData}{experimental clearance value}

\item{typeSystem}{microsomes or hepatocytes}

\item{fu_invitro}{Fraction unbound in the in vitro hepatic system. if not known code will calculate it}

\item{empirical_scalar}{this is an option to include an extra empirical correction factor. Currently we are considering the scale factor of Wood 2017}

\item{tissue}{tissue of interest, since scaling factors are dependent on the tissue, will default to liver}

\item{species}{values can human and rat for now, defaulting to human}

\item{volMedium_mL}{volume of medium in the well (in mL)}

\item{REF}{relative expression or activity factor}

\item{cMicro_mgml}{concentration of microsome protein (in mg/mL)}

\item{cCells_Mml}{concentration of hepatocytes used (in million cells/mL)}
}
\value{
Specific clearance parameter (/min) to plug in PK-Sim
}
\description{
IVIVE for clearance based on different type of values
}
\examples{
# example hepatocytes
IVIVE_clearance(typeValue="in vitro clearance parameter",typeSystem="hepatocytes",species="human",units="mL/minutes/millioncells",
expData=18.27,fu_invitro=0.5,cCells_Mml=0.5,empirical_scalar="No")

# example microsomes
IVIVE_clearance(typeValue="in vitro clearance parameter",typeSystem="microsomes",units="L/minutes/mg protein",
                expData=18.27,fu_invitro=0.5,cMicro_mgml=0.5,volMedium_mL=0.5,empirical_scalar="No")

}
