% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVE_MM.R
\name{IVIVE_MM}
\alias{IVIVE_MM}
\title{Title}
\usage{
IVIVE_MM(
  typeSystem,
  fu_invitro = 1,
  vmax,
  km_micromolar,
  tissue = "liver",
  species = "human",
  REF = 1
)
}
\arguments{
\item{typeSystem}{if hepatocytes or microsomes}

\item{fu_invitro}{value of fractionunbound in vitro, the default is 1}

\item{vmax}{as umol/min/million hepatocytes or umol/min/mg microsomal protein}

\item{km_micromolar}{}

\item{tissue}{liver, brain, lung, kidney, gonads and gut, default is liver}

\item{species}{human, rat or dog, default human}

\item{REF}{relative expression or activity factor, default is 1.
To use this option the reference concentration of
of the enzyme of interest in pksim needs to be 1 uM}
}
\value{
Vmax in umol/min/L and Km_unb in uM
}
\description{
Title
}
\examples{

IVIVE_MM (typeSystem="hepatocytes",vmax=2,km_micromolar=1,tissue="liver",species="human",REF=1)
IVIVE_MM (typeSystem="microsomes",fu_invitro=0.2,vmax=2,km_micromolar=1)
}
