% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fu-krumpholz.R
\name{list_krumpholz_compounds}
\alias{list_krumpholz_compounds}
\title{List compounds in the Krumpholz et al. database}
\usage{
list_krumpholz_compounds(system = "microsomes")
}
\arguments{
\item{system}{in vitro system: "microsomes", "hepatocytes", "plasma" or "recombinant CYPs"}
}
\value{
sorted character vector with the compound names available for that system
}
\description{
List compounds in the Krumpholz et al. database
}
\examples{
head(list_krumpholz_compounds("hepatocytes"))
}
